### 2017-06-20 #290 @Kevin-Mattheus-Moerman
I am quite swamped at the moment but will check briefly to see if I can find a reviewer here at MIT. Can the authors recommend any reviewers at this point?

In the mean time (I gave this a super quick glance) could the authors expand the [readme](https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/blob/master/README.md) to leave a few more bread crumbs for people to follow when they view the project. Here is an example of a more complete [readme](https://github.com/gibbonCode/GIBBON/blob/master/README.md). Perhaps directly link to the [wiki](https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/wiki/How-to-Set-Up-the-Project) from the readme.md (rather than the GitHub page which is where readers already are). Can you also link to an actual license file in your repository? I recommend adding a graphical example / screen-shot if possible.
