### 2017-06-22 #290 @Kevin-Mattheus-Moerman
I just found this: https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/wiki/Running-a-Simulation
That link gives a much better impression on what the software does than the readme or paper. As I said please expand the readme and the paper on what the software is for and try to include graphical examples on how it is used.
I'm waiting to hear back from potential reviewers.
