### 2017-06-28 #290 @Kevin-Mattheus-Moerman
@AlexisDrogoul @rinde @eclab @alexlin0 at [JOSS](http://joss.theoj.org/) we are looking for reviewers for this submission. If [this work](https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/) looks interesting to you and you are willing to [review](http://joss.theoj.org/about#reviewer_guidelines) for JOSS please let me know.
