### 2017-07-26 #290 @kavonszadkowski
Hi Kevin,

sorry to decline, but I don't think this fits my area of expertise well.

Cheers,

Kai

> @Denz1994 @samreid @mattpen @agrignard @AlexisDrogoul @rinde @eclab @alexlin0 @kavonszadkowski at [JOSS](http://joss.theoj.org/) we are looking for reviewers for this submission. If [this work](https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/) looks interesting to you and you are willing to [review](http://joss.theoj.org/about#reviewer_guidelines) for JOSS please let me know.
>

--
  Kai von Szadkowski (M.Sc.)
  Space Robotics

  Universität Bremen
  FB 3 - Mathematik und Informatik
  AG Robotik
  Robert-Hooke-Straße 1
  28359 Bremen, Germany

  Zentrale: +49 421 178 45-6611

  Besuchsadresse der Nebengeschäftstelle:
  Robert-Hooke-Straße 5
  28359 Bremen, Germany

  Tel.:    +49 421 178 45-6646
  Empfang: +49 421 178 45-6600
  Fax:     +49 421 178 45-4150
  E-Mail:   kai.von-szadkowski@uni-bremen.de

  Weitere Informationen: http://www.informatik.uni-bremen.de/robotik
