### 2017-07-27 #290 @rinde
Hi Kevin,

Thanks for the invite, but I unfortunately don't have the time to do a review.

Cheers,
Rinde
