### 2017-09-24 #290 @arfon
👋 friendly reminder @zmimlitz to reply to @Kevin-Mattheus-Moerman's [comment](https://github.com/openjournals/joss-reviews/issues/290#issuecomment-318907992)
