### 2017-10-20 #434 @whedon
[REVIEW]: Simulated Physics and Environment for Autonomous Risk Studies (SPEARS)
**Submitting author:** @zmimlitz (<a href="http://orcid.org/0000-0001-5431-8362">Zac Mimlitz</a>)
**Repository:** <a href="https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/" target ="_blank">https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/</a>
**Version:** 3.0.5
