### 2017-10-20 #290 @Kevin-Mattheus-Moerman
@agrignard thanks for reviewing with JOSS!

@zmimlitz We've started the review process over at #434. I'll close this pre-review issue.
