### 2017-10-20 Van-Bossuyt-Research-Lab/SPEARS_Simulator#35 @agrignard
Documentation update
Hi,

I am a reviewer of your submission to JOSS.

I am trying to install your librairie from Eclipse (Version: Mars.2 Release (4.5.2))

If I follow the instruction then when I am trying to create the java application I gave the name Run and then what? I have this message error saying

Is there anything I am missing?
