### 2017-11-07 Van-Bossuyt-Research-Lab/SPEARS_Simulator#35 @agrignard
Any update on that? I am a bit stop in the process of reviewing your software in JOSS
