### 2017-11-15 #434 @agrignard
Actually yes, I was not able to run the software following the github instruction, i created an issue but I am still waiting for there answer

https://github.com/Van-Bossuyt-Research-Lab/SPEARS_Simulator/issues/35

So I am kind of stuck in the review process, what should I do ?
