### 2018-01-08 #434 @arfon
> @arfon can you e-mail the authors a final reminder? If this review issue remains stale I suggest we retract this submission.

Done (and agreed).
