### 2019-02-15 #434 @douglasvb
Just saw these comments a year too late.  For whatever reason, I never was notified.  Is there interest from JOSS to have us fix the error and re-open the review?
