### 2019-02-15 #434 @douglasvb
@Kevin-Mattheus-Moerman I'm checking with @zmimlitz to see about getting the code fixed.  Perhaps it would be appropriate for us to fix the issue first before we have you reopen the review.

I believe the issue was the instructions for compiling the code wasn't working?  Was there anything else that we should address?
