### 2019-02-15 #434 @arfon
> I believe the issue was the instructions for compiling the code wasn't working? Was there anything else that we should address?

There could potentially be much more work to do here as @agrignard looks to have only just started their review.

@Kevin-Mattheus-Moerman - we can reopen this issue yes, but I think we should get a 👍 from @agrignard that they are still willing to review this submission. If not, @douglasvb I think you need to resubmit.
