### 2019-02-15 #434 @douglasvb
If you're willing to bear with us a little, I think we should be able to respond to issues as they come up in the review process.  I'm working the installation instructions issue now.

Again, apologies that we dropped the ball on the initial review.
